add wasm stuff
