+++ Makefile
@@ -1,7 +1,23 @@
+GOROOT=$(shell go env GOROOT)
+GOMOD=$(shell test -f "go.work" && echo "readonly" || echo "vendor")
+LDFLAGS=-s -w
+
 .PHONY: build
 build:
-	go build -v -o bin/wof-format ./cmd
+	@make cli
+	@make wasmjs
+
+cli:
+	go build -mod $(GOMOD) -ldflags="$(LDFLAGS)" \
+		-o bin/wof-format \
+		./cmd/wof-format/main.go
+
+wasmjs:
+	GOOS=js GOARCH=wasm \
+		go build -mod $(GOMOD) -ldflags="$(LDFLAGS)" -tags wasmjs \
+		-o static/wasm/wof_format.wasm \
+		cmd/wof-format-wasm/main.go
 
 .PHONY: test
 test:
-	go test -v ./test/...
+	go test -v ./...
